+++ Editor/CurveEditUtility.cs
@@ -614,20 +614,7 @@ namespace UnityEditor.Timeline
         }
 
         // Creates an opposing blend curve that matches the given curve to make sure the result is normalized
-        public static AnimationCurve CreateMatchingCurve(AnimationCurve curve)
-        {
-            Keyframe[] keys = curve.keys;
-
-            for (var i = 0; i != keys.Length; i++)
-            {
-                if (!Single.IsPositiveInfinity(keys[i].inTangent))
-                    keys[i].inTangent = -keys[i].inTangent;
-                if (!Single.IsPositiveInfinity(keys[i].outTangent))
-                    keys[i].outTangent = -keys[i].outTangent;
-                keys[i].value = 1.0f - keys[i].value;
-            }
-            return new AnimationCurve(keys);
-        }
+        public static AnimationCurve CreateMatchingCurve(AnimationCurve curve) => UnityEngine.Timeline.CurveEditUtility.CreateMatchingCurve(curve);
 
         // Sanitizes the keys on an animation to force the property to be normalized
         public static Keyframe[] SanitizeCurveKeys(Keyframe[] keys, bool easeIn)